downtime_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
celery==5.4.0
pika==1.3.2
PyJWT==2.9.0
